flask/requirements.txt
flask
flask_restful
flask_cors
pixellib
opencv-python
pillow
pycocotools
scikit-image
imutils
### mac ###
tensorflow-macos
tensorflow-metal
### win ###
#tensorflow
#tensorflow-gpu
